Prototype: Use Firestore instead of Datastore
Drive-by fix: simplify file structure for tests.
