--- 3-binary-data/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "nodejs-getting-started",
-  "description": "End to end sample for running Node.js applications on Google Cloud Platform",
-  "version": "1.0.0",
-  "license": "Apache-2.0",
-  "author": "Google Inc.",
-  "engines": {
-    "node": ">=8"
-  },
-  "repository": "https://github.com/GoogleCloudPlatform/nodejs-getting-started",
-  "main": "app.js",
-  "private": true,
-  "cloud-repo-tools": {
-    "requiresKeyFile": true,
-    "requiresProjectId": true,
-    "test": {
-      "app": {
-        "args": [
-          "app.js"
-        ],
-        "msg": "Bookshelf - Node.js on Google Cloud Platform"
-      }
-    }
-  },
-  "scripts": {
-    "start": "node app.js",
-    "test": "repo-tools test app && ava -t 30s --tap test/*.test.js",
-    "e2e": "repo-tools test deploy && E2E_TESTS=True ava -t 30s --tap test/*.test.js"
-  },
-  "dependencies": {
-    "@google-cloud/datastore": "^4.0.0",
-    "@google-cloud/storage": "^1.7.0",
-    "body-parser": "^1.18.3",
-    "express": "^4.16.4",
-    "lodash": "^4.17.11",
-    "multer": "^1.4.1",
-    "nconf": "^0.10.0",
-    "prompt": "^1.0.0",
-    "pug": "^2.0.3",
-    "uglify-js": "^3.4.9"
-  },
-  "devDependencies": {
-    "@google-cloud/nodejs-repo-tools": "^3.0.0",
-    "ava": "^0.25.0",
-    "proxyquire": "^2.1.0",
-    "sinon": "^7.0.0"
-  }
-}
