+++ package.json
@@ -16,9 +16,8 @@
     "test": "npm run lint && npm run mocha"
   },
   "dependencies": {
-    "@google-cloud/connect-datastore": "^3.0.0",
-    "@google-cloud/datastore": "^4.0.0",
-    "@google-cloud/debug-agent": "^4.0.0",
+    "@google-cloud/debug-agent": "^3.0.0",
+    "@google-cloud/firestore": "^1.0.0",
     "@google-cloud/logging-winston": "^0.11.0",
     "@google-cloud/pubsub": "^0.30.0",
     "@google-cloud/storage": "^1.6.0",
